recreate the file structure, for the future pc/mobile.
