+++ lib/components/common/left_drawer/chapter_listview.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
 import 'package:redux/redux.dart';
-import '../../redux/action/text_action.dart';
-import '../../redux/app_state/state.dart';
+import '../../../redux/action/text_action.dart';
+import '../../../redux/app_state/state.dart';
 
 class ChapterListView extends StatefulWidget {
   final String bookName;
