+++ lib/components/common/right_drawer/right_drawer.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
 import 'package:redux/redux.dart';
-import 'package:writing_writer/components/right_drawer/set_listview.dart';
-import '../../redux/app_state/state.dart';
-import '../../server/style/StyleBase.dart';
+import 'package:writing_writer/components/common/right_drawer/set_listview.dart';
+import '../../../redux/app_state/state.dart';
+import '../../../server/style/StyleBase.dart';
 import '../toast_dialog.dart';
 import '../transparent_checkbox.dart';
