+++ lib/components/common/right_drawer/set_listview.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
 import 'package:redux/redux.dart';
-import 'package:writing_writer/components/right_drawer/settings_listview.dart';
-import '../../redux/app_state/state.dart';
-import '../../server/parser/Parser.dart';
+import 'package:writing_writer/components/common/right_drawer/settings_listview.dart';
+import '../../../redux/app_state/state.dart';
+import '../../../server/parser/Parser.dart';
 import '../toast_dialog.dart';
 import '../transparent_icon_button.dart';
