+++ lib/components/common/right_drawer/settings_listview.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
 import 'package:redux/redux.dart';
-import '../../redux/action/set_action.dart';
-import '../../redux/app_state/state.dart';
+
+import '../../../redux/action/set_action.dart';
+import '../../../redux/app_state/state.dart';
 
 class SettingsListView extends StatefulWidget {
   final String setName;
