+++ lib/entry/mobile/mobile_main.dart
@@ -1,0 +1,33 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_redux/flutter_redux.dart';
+import 'package:redux/redux.dart';
+import 'package:writing_writer/redux/app_state/state.dart';
+import 'package:writing_writer/router/pc/pc_router.dart';
+
+class MobileApp extends StatelessWidget {
+  MobileApp({Key? key}) : super(key: key);
+
+  final store = Store<AppState>(appReducer, initialState: AppState.initialState());
+
+  @override
+  Widget build(BuildContext context) {
+    /// 使用StoreProvider 包裹根元素，使其提供store
+    return StoreProvider(
+      store: store,
+      /// 为了能直接在child使用store，我们这里要继续包裹一层StoreBuilder
+      child: StoreBuilder<AppState>(
+        builder: (context, store) {
+          return MaterialApp(
+              title: 'Writing Writer',
+              theme: ThemeData(
+                brightness: Brightness.dark,
+              ),
+              debugShowCheckedModeBanner: false,
+              initialRoute: '/',
+              onGenerateRoute: onGenerateRoute
+          );
+        },
+      ),
+    );
+  }
+}
