--- lib/unusedNow/parser_action.dart
@@ -1,25 +1,0 @@
-import 'package:redux/redux.dart';
-
-import 'parser_model.dart';
-
-class SetParserDataAction{
-  /// 当前设定集解析加入展示
-  Map<String, Set<String>> currentParser;
-
-  SetParserDataAction({
-    required this.currentParser,
-  });
-
-  static ParserModel setParser(ParserModel parserModel, SetParserDataAction action) {
-    parserModel.currentParser = action.currentParser;
-    return parserModel;
-  }
-
-}
-
-/*
- * 绑定Action与动作
- */
-final parserReducer = combineReducers<ParserModel>([
-  TypedReducer<ParserModel, SetParserDataAction>(SetParserDataAction.setParser),
-]);
