--- lib/unusedNow/parser_model.dart
@@ -1,8 +1,0 @@
-class ParserModel{
-  /// 当前设定集解析加入展示
-  Map<String, Set<String>> currentParser;
-
-  ParserModel({
-    required this.currentParser,
-  });
-}
